requirements.txt
pyautogui>=0.9.54
pytesseract>=0.3.10
pillow>=10.0.0
PyQt6>=6.4.0
numpy>=1.24.0